change the data type of id to Integer and added postmapping for getting the emp record using id.Correct the data.sql file

+++ src/main/java/com/example/demo/employeeTable/repository/EmployeeRepository.java
@@ -2,10 +2,10 @@ package com.example.demo.employeeTable.repository;
 
 import com.example.demo.employeeTable.entity.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.web.bind.annotation.GetMapping;
+
 
 import java.util.List;
-public interface EmployeeRepository extends JpaRepository<Employee,Float> {
+public interface EmployeeRepository extends JpaRepository<Employee,Integer> {
 
     List<Employee> findAll();
 }
